% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s01_main.R
\name{.plot_duo}
\alias{.plot_duo}
\title{.plot_duo}
\usage{
.plot_duo(
  dat = irr,
  X = "CD",
  Y = "value",
  Z = "대안",
  W = "임대기간",
  alt = c(비교 = "비교", 현재 = "현재"),
  .w = c("10년", "10년"),
  xlbl = "CD금리(\%)",
  ylbl = "value",
  xcri = c(`3.5` = 3.5),
  ycri = c(0),
  a_ratio = 3/4,
  .guide = T,
  .fntSz = 16,
  .percent = T,
  .digits = 3,
  .ylim = -9,
  .yAxis = 0
)
}
\arguments{
\item{dat}{데이터 프레임}

\item{X}{x축 변수명}

\item{Y}{y축 변수명}

\item{Z}{비교 변수명}

\item{W}{자료 분리 변수명}

\item{alt}{비교군(범례) 목록}

\item{.w}{비교군(범례) 목록에 대응하는 자료 분리 변수값}

\item{xlbl}{x축 레이블}

\item{ylbl}{y축 레이블}

\item{xcri}{x축 기준값}

\item{ycri}{y축 기준값}

\item{a_ratio}{세로:가로 비율}

\item{.guide}{범례 표시 여부}

\item{.fntSz}{폰트 크기}

\item{.percent}{y축 백분율 표시 여부}

\item{.digits}{y축 소수점 자리수}

\item{.ylim}{y축 한계값}

\item{.yAxis}{y축 교차값}
}
\value{
그래프와 t.test 결과
}
\description{
여러 데이터를 비교하는 그래프를 그리는 함수
}
