% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s01_main.R
\name{monthly_to_daily_simple}
\alias{monthly_to_daily_simple}
\title{monthly_to_daily_simple}
\usage{
monthly_to_daily_simple(posix_dates, values)
}
\arguments{
\item{posix_dates}{POSIXct 클래스의 날짜 벡터}

\item{values}{변환할 값 벡터}
}
\value{
변환된 일별 데이터 프레임
}
\description{
월별 데이터를 일별 데이터로 변환하는 함수
}
