% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s01_main.R
\name{readKOSIS}
\alias{readKOSIS}
\title{readKOSIS}
\usage{
readKOSIS(url)
}
\arguments{
\item{url}{URL of KOSIS statistic table}
}
\value{
Data frame of KOSIS statistic table
}
\description{
Read KOSIS statistic table into data frame using URL
}
